+++ CHANGELOG.md
@@ -66,6 +66,7 @@
 #### Reverse proxying
 
 - **Reverse proxy stall fix**: fixed stalls on some connections caused by aborted connections when the reverse proxy was waiting for a response from an HTTP/1.x upstream server.
+- **Circuit breaker metric fix**: fixed missing circuit breaker metrics when circuit breaker is enabled but not yet opened, so that the `ferron.proxy.circuit_breaker.state` metric is always emitted.
 
 ## Ferron 3.0.0-rc.6